refactor: change `__init__` API; fix: use logger only when provided. feat: now a frameserver may manage two different ring buffer with the same flow control. TODO: Useful? How to test?

+++ frameserver_v2_types.py
@@ -11,8 +11,8 @@ class FrameTicket:
     """
     FrameTicket is issued by FrameServer and use absolute index to identify frame batches.
     """
-    head_id: int    # The global index maintained by FrameServer
-    size: int       # The continuous frames to get started from head_id.
+    head_id: np.int64  # The global index maintained by FrameServer
+    size: int          # The continuous frames to get started from head_id.
 
 class TicketExpireException(Exception):
     """Ticket is expired."""
